break code into separate files, start refactoring of subgraph entity
