+++ compiler/luci-codegen/src/Driver.cpp
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2021 Samsung Electronics Co., Ltd. All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include "LuciCodegen.h"
 #include "luci/Importer.h"
 #include "luci/CircleExporter.h"
@@ -22,7 +38,7 @@ int main(int argc, char **argv)
   luci_codegen::Options options;
   // set options if needed
   luci_codegen::LuciCodegen codegen(options);
-  codegen.process(*luci_module);
+  codegen.process_module(*luci_module);
   codegen.emit_code(output_package_name);
 
   luci::CircleExporter exporter;
